refactor out unused calc
